COMP: Prefer size_t for memory indicies access
Variables that identify memory locations (either size, or access)
should use size_t.

http://stackoverflow.com/questions/1951519/when-should-i-use-stdsize-t

A good rule of thumb is for anything that you need to compare in the loop condition against something that is naturally a std::size_t itself.

std::size_t is the type of any sizeof expression and as is guaranteed to be able to express the maximum size of any object (including any array) in C++. By extension it is also guaranteed to be big enough for any array index so it is a natural type for a loop by index over an array.

If you are just counting up to a number then it may be more natural to use either the type of the variable that holds that number or an int or unsigned int (if large enough) as these should be a natural size for the machine.

+++ core/vsl/vsl_basic_xml_element.h
@@ -44,7 +44,7 @@ class vsl_basic_xml_element
   void add_attribute(std::string attr_name, float value) { add_attribute(attr_name, (double)value); }
   void add_attribute(std::string attr_name, long value);
   void add_attribute(std::string attr_name, int value) { add_attribute(attr_name, (long)value); }
-  void add_attribute(std::string attr_name, unsigned long value) { add_attribute(attr_name, (long)value); }
+  void add_attribute(std::string attr_name, size_t value) { add_attribute(attr_name, (long)value); }
   void add_attribute(std::string attr_name, unsigned int value) { add_attribute(attr_name, (long)value); }
 
   void add_attribute_list(std::vector<std::pair<std::string, std::string> > attrs);
